requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
python-pptx
pymupdf
